immo_pulse/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
coreapi==2.3.3
coreschema==0.0.4
configparser==5.2.0
charset-normalizer==2.0.7
cryptography==35.0.0
Django==3.2.9
django-rest-swagger==2.2.0
djangorestframework==3.12.2
django-sslserver==0.22
django-filter==2.4.0
django-enum-choices==2.1.3
django-health-check==3.16.5
idna==3.3
isodate==0.6.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
msrest==0.6.19
numpy==1.21.4
oauthlib==3.1.1
pip==21.3.1
pandas==1.3.4
psycopg2-binary==2.9.2
pycparser==2.20
pyaes==1.6.1
pyasn1==0.4.8
python-dateutil==2.8.2
pytz==2020.5
pycryptodome==3.12.0
PyJWT==2.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sqlparse==0.4.2
urllib3==1.26.7
